rebase master and fix error hover shadow
