+++ src/form/style/index.less
@@ -501,15 +501,6 @@ form {
   // input
   .@{ant-prefix}-input-affix-wrapper {
     border-color: @warning-color;
-    box-shadow: @input-outline-offset @outline-blur-size @outline-width
-      fade(@warning-color, @outline-fade);
-    .@{ant-prefix}-input {
-      &:hover,
-      &:focus {
-        border: none;
-        box-shadow: none;
-      }
-    }
   }
 }
 
@@ -614,15 +605,6 @@ form {
   // input
   .@{ant-prefix}-input-affix-wrapper {
     border-color: @error-color;
-    box-shadow: @input-outline-offset @outline-blur-size @outline-width
-      fade(@error-color, @outline-fade);
-    .@{ant-prefix}-input {
-      &:hover,
-      &:focus {
-        border: none;
-        box-shadow: none;
-      }
-    }
   }
 }
